Refactor backends into directory
