+++ src/scmdata/__init__.py
@@ -16,3 +16,4 @@ except Exception:  # pylint: disable=broad-except  # pragma: no cover
     __version__ = "unknown"
 
 from scmdata.run import ScmRun, run_append  # noqa: F401, E402
+from scmdata.database import ScmDatabase  # noqa: F401, E402
