+++ src/scmdata/database/database.py
@@ -9,313 +9,16 @@ Subsets of data are also able to be read without having to load all the data
 and then filter. For example, one could save model results from a number of different
 climate models and then load just the ``Surface Temperature`` data for all models.
 """
-import glob
-import itertools
 import os
 import os.path
-import pathlib
-from abc import ABC, abstractmethod
 
 import pandas as pd
 import six
 import tqdm.autonotebook as tqdman
 
-from scmdata import ScmRun, run_append
-
-
-def ensure_dir_exists(fp):
-    """
-    Ensure directory exists
-
-    Parameters
-    ----------
-    fp : str
-        Filepath of which to ensure the directory exists
-    """
-    dir_to_check = os.path.dirname(fp)
-    if not os.path.isdir(dir_to_check):
-        try:
-            os.makedirs(dir_to_check)
-        except OSError:  # pragma: no cover
-            # Prevent race conditions if multiple threads attempt to create dir at same time
-            if not os.path.exists(dir_to_check):
-                raise
-
-
-def _check_is_subdir(root, d):
-    root_path = pathlib.Path(root).resolve()
-    out_path = pathlib.Path(d).resolve()
-
-    is_subdir = root_path in out_path.parents
-    # Sanity check that we never mangle anything outside of the root dir
-    if not is_subdir:  # pragma: no cover
-        raise AssertionError("{} not in {}".format(d, root))
-
-
-def _get_safe_filename(inp, include_glob=False):
-    def safe_char(c):
-        accepted_chars = "-_."
-        if include_glob:
-            accepted_chars = accepted_chars + "*"
-
-        if c.isalnum() or c in accepted_chars or c == os.sep:
-            return c
-
-        return "-"
-
-    return "".join(safe_char(c) for c in inp)
-
-
-class DatabaseBackend(ABC):
-    """
-    Abstract backend for serialising/deserialising data
-
-    Data is stored as objects represented by keys. These keys can be used later
-    to load data.
-    """
-
-    def __init__(self, **kwargs):
-        self.kwargs = kwargs
-
-    @abstractmethod
-    def save(self, sr):
-        """
-        Save data
-
-        Parameters
-        ----------
-        sr: scmdata.ScmRun
-
-        Returns
-        -------
-        str
-            Key where the data is stored
-        """
-        pass
-
-    @abstractmethod
-    def load(self, key):
-        """
-        Load data at a given key
-
-        Parameters
-        ----------
-        key : str
-            Key to load
-
-        Returns
-        -------
-        scmdata.ScmRun
-        """
-        pass
-
-    def delete(self, key):
-        """
-        Delete a given key
-
-        Parameters
-        ----------
-        key: str
-        """
-        pass
-
-    @abstractmethod
-    def get(self, filters):
-        """
-        Get all matching keys for a given filter
-
-        Parameters
-        ----------
-        filters: dict of str
-            String filters
-            If a level is missing then all values are fetched
-
-        Returns
-        -------
-        list of str
-            Each item is a key which may contain data which is of interest
-        """
-        pass
-
-
-class NetCDFBackend(DatabaseBackend):
-    """
-    On-disk database handler for outputs from SCMs
-
-    Data is split into groups as specified by :attr:`levels`. This allows for fast
-    reading and writing of new subsets of data when a single output file is no longer
-    performant or data cannot all fit in memory.
-    """
-
-    def get_key(self, sr):
-        """
-        Get key where the data will be stored
-
-        The key is the root directory joined with the other information provided. The filepath
-        is also cleaned to remove spaces and special characters.
-
-        Parameters
-        ----------
-        sr : :class:`scmdata.ScmRun`
-            Data to save
-
-        Raises
-        ------
-        ValueError
-            If non-unique metadata is found for each of :attr:`self.kwargs["levels"]`
-
-            If any metadata end with '.'
-
-        KeyError
-            If missing metadata is found for each of :attr:`self.kwargs["levels"]`
-
-        Returns
-        -------
-        str
-            Path in which to save the data without spaces or special characters
-        """
-        levels = {
-            database_level: sr.get_unique_meta(
-                database_level, no_duplicates=True
-            ).replace(os.sep, "_")
-            for database_level in self.kwargs["levels"]
-        }
-
-        # Windows does not support directories or filenames which end in a '.'
-        if any([level.endswith(".") for level in levels.values()]):
-            raise ValueError("Metadata cannot end in a '.'")
-
-        return self._get_out_filepath(**levels)
-
-    def _get_out_filepath(self, **data_levels):
-        out_levels = []
-        for database_level in self.kwargs["levels"]:
-            if database_level not in data_levels:  # pragma: no cover # emergency valve
-                raise KeyError("expected level: {}".format(database_level))
-            out_levels.append(str(data_levels[database_level]))
-
-        out_path = os.path.join(*out_levels)
-        out_fname = "__".join(out_levels) + ".nc"
-        out_fname = _get_safe_filename(os.path.join(out_path, out_fname))
-        out_fname = os.path.join(self.kwargs["root_dir"], out_fname)
-
-        _check_is_subdir(self.kwargs["root_dir"], out_fname)
-
-        return out_fname
-
-    def save(self, sr):
-        """
-        Save a ScmRun to the database
-
-        The dataset should not contain any duplicate metadata for the
-        database levels
-
-        Parameters
-        ----------
-        sr : :class:`scmdata.ScmRun`
-            Data to save
-
-        Raises
-        ------
-        ValueError
-            If duplicate metadata are present for the requested database levels
-
-        KeyError
-            If metadata for the requested database levels are not found
-
-        Returns
-        -------
-        str
-            Key where the data is saved
-        """
-        key = self.get_key(sr)
-
-        ensure_dir_exists(key)
-        if os.path.exists(key):
-            existing_run = ScmRun.from_nc(key)
-
-            sr = run_append([existing_run, sr])
-
-        # Check for required extra dimensions
-        dimensions = self.kwargs.get("dimensions", None)
-        if not dimensions:
-            nunique_meta_vals = sr.meta.nunique()
-            dimensions = nunique_meta_vals[nunique_meta_vals > 1].index.tolist()
-        sr.to_nc(key, dimensions=dimensions)
-        return key
-
-    def load(self, key):
-        """
-
-        Parameters
-        ----------
-        key: str
-
-        Returns
-        -------
-        :class:`scmdata.ScmRun`
-
-        """
-        return ScmRun.from_nc(key)
-
-    def delete(self, key):
-        """
-        Delete a key
-
-        Parameters
-        ----------
-        key: str
-
-        """
-        os.remove(key)
-
-    def get(self, filters):
-        """
-        Get all matching objects for a given filter
-
-        Parameters
-        ----------
-        filters: dict of str
-            String filters
-            If a level is missing then all values are fetched
-
-        Returns
-        -------
-        list of str
-        """
-        level_options = []
-        for level in self.kwargs["levels"]:
-            level_values = filters.get(level, ["*"])
-            if isinstance(level_values, str):
-                level_values = [level_values]
-
-            level_options.append(level_values)
-
-        # AND logic across levels, OR logic within levels
-        level_options_product = itertools.product(*level_options)
-        globs_to_check = [
-            _get_safe_filename(os.path.join(*levels, "*.nc"), include_glob=True)
-            for levels in level_options_product
-        ]
-
-        load_files = [
-            v
-            for vlist in [
-                glob.glob(os.path.join(self.kwargs["root_dir"], g), recursive=True)
-                for g in globs_to_check
-            ]
-            for v in vlist
-        ]
-
-        return load_files
-
-
-"""
-Loaded backends for ScmDatabase
-
-Additional backends should be based upon :class:`DatabaseBackend`
-"""
-backend_classes = {"netcdf": NetCDFBackend}
+from scmdata import run_append
+from scmdata.database.backends import backend_classes, DatabaseBackend
+from scmdata.database.utils import _check_is_subdir
 
 
 class ScmDatabase:
